replace retrieveApplicantsFromList in main file
