+++ app/javascript/react/pages/ApplicantsUpload.jsx
@@ -1,15 +1,25 @@
 import React, { useState, useReducer } from "react";
 
+import * as XLSX from "xlsx";
 import FileHandler from "../components/FileHandler";
 import ApplicantList from "../components/ApplicantList";
 import EnrichWithContactFile from "../components/EnrichWithContactFile";
 
-import { parameterizeObjectValues } from "../../lib/parameterize";
-import retrieveApplicantsFromList from "../lib/retrieveApplicantsFromList";
+import getHeaderNames from "../lib/getHeaderNames";
+import checkColumnNames from "../lib/checkColumnNames";
+import displayMissingColumnsWarning from "../lib/displayMissingColumnsWarning";
 import retrieveUpToDateApplicants from "../lib/retrieveUpToDateApplicants";
 import updateApplicantContacts from "../lib/updateApplicantContacts";
 import retrieveContactsData from "../lib/retrieveContactsData";
 import { initReducer, reducerFactory } from "../../lib/reducers";
+import { excelDateToString } from "../../lib/datesHelper";
+import {
+  parameterizeObjectKeys,
+  parameterizeObjectValues,
+  parameterizeArray,
+} from "../../lib/parameterize";
+
+import Applicant from "../models/Applicant";
 
 const reducer = reducerFactory("Expérimentation RSA");
 
@@ -35,18 +45,79 @@ export default function ApplicantsUpload({ organisation, configuration, departme
       : `/organisations/${organisation.id}/applicants`;
   };
 
+  const retrieveApplicantsFromList = async (file) => {
+    const applicantsFromList = [];
+
+    await new Promise((resolve) => {
+      const reader = new FileReader();
+      reader.onload = function (event) {
+        const sheetName = configuration.sheet_name;
+        const workbook = XLSX.read(event.target.result, { type: "binary" });
+        const sheet = workbook.Sheets[sheetName] || workbook.Sheets[workbook.SheetNames[0]];
+        const headerNames = getHeaderNames(sheet);
+        const missingColumnNames = checkColumnNames(
+          columnNames,
+          configuration,
+          parameterizeArray(headerNames)
+        );
+        if (missingColumnNames.length > 0) {
+          displayMissingColumnsWarning(missingColumnNames);
+        } else {
+          let rows = XLSX.utils.sheet_to_row_object_array(sheet);
+          rows = rows.map((row) => parameterizeObjectKeys(row));
+          rows.forEach((row) => {
+            const applicant = new Applicant(
+              {
+                lastName: row[parameterizedColumnNames.last_name],
+                firstName: row[parameterizedColumnNames.first_name],
+                affiliationNumber: row[parameterizedColumnNames.affiliation_number],
+                role: row[parameterizedColumnNames.role],
+                title: row[parameterizedColumnNames.title],
+                address: parameterizedColumnNames.address && row[parameterizedColumnNames.address],
+                fullAddress:
+                  parameterizedColumnNames.full_address &&
+                  row[parameterizedColumnNames.full_address],
+                email: parameterizedColumnNames.email && row[parameterizedColumnNames.email],
+                birthDate:
+                  parameterizedColumnNames.birth_date &&
+                  row[parameterizedColumnNames.birth_date] &&
+                  excelDateToString(row[parameterizedColumnNames.birth_date]),
+                city: parameterizedColumnNames.city && row[parameterizedColumnNames.city],
+                postalCode:
+                  parameterizedColumnNames.postal_code && row[parameterizedColumnNames.postal_code],
+                phoneNumber:
+                  parameterizedColumnNames.phone_number &&
+                  row[parameterizedColumnNames.phone_number],
+                birthName:
+                  parameterizedColumnNames.birth_name && row[parameterizedColumnNames.birth_name],
+                departmentInternalId:
+                  parameterizedColumnNames.department_internal_id &&
+                  row[parameterizedColumnNames.department_internal_id],
+                rightsOpeningDate:
+                  parameterizedColumnNames.rights_opening_date &&
+                  row[parameterizedColumnNames.rights_opening_date] &&
+                  excelDateToString(row[parameterizedColumnNames.rights_opening_date]),
+              },
+              department,
+              organisation,
+              configuration
+            );
+            applicantsFromList.push(applicant);
+          });
+        }
+        resolve();
+      };
+      reader.readAsBinaryString(file);
+    });
+
+    return applicantsFromList.reverse();
+  };
+
   const handleApplicantsFile = async (file) => {
     setFileSize(file.size);
 
     dispatchApplicants({ type: "reset" });
-    const applicantsFromList = await retrieveApplicantsFromList(
-      file,
-      organisation,
-      department,
-      configuration,
-      columnNames,
-      parameterizedColumnNames
-    );
+    const applicantsFromList = await retrieveApplicantsFromList(file);
     if (applicantsFromList.length === 0) return;
 
     const upToDateApplicants = await retrieveUpToDateApplicants(applicantsFromList);
